Move version-neutral management code out of v1/ into shared modules
v1 will eventually be abandoned, so the layout now enforces one invariant:
nothing under management/v2/ may import from management/v1/. That makes v1
removable with `rm -rf management/v1/` plus deletion of the back-compat shims.

To get there, version-neutral *mechanics* move up into the top-level
management/ package and the version packages keep only version-specific
*vocabulary*. This inverts the old arrangement, where the top-level modules
were thin re-export shims over v1 and v2 subclassed v1 directly.

Live probing (read-only GETs against api.singlestore.com) confirmed which
routes are actually version-neutral:

  files/fs/{space}      200 at both     -> management/files.py
  jobs                  405 at both     -> management/job.py
  billing/usage         400 at both     -> management/billing{,_usage}.py
  organizations/current 200 at both     -> management/organization.py
  secrets               200 at both     -> management/organization.py
  regions               200 at both     -> management/region.py
  regions/sharedtier    v1 only         -> v2 RegionManager raises
  models                v1 only         -> v2 InferenceAPIManager raises

Stage moves to management/stage.py behind the _fs_path() hook added earlier,
since at v2 it hangs off the cluster resource instead of being top-level.
Billing and Organizations move out of v1/workspace.py to management/billing.py
and management/organization.py.

Two spots needed a hook rather than a straight move:

  - JobsManager grew _resolve_target() plus three _*_target_type class
    attributes. The jobs routes are unchanged at v2; only the
    targetConfig.targetType vocabulary is, and 'Cluster' means different
    things at each version, so TargetType now holds the union of both
    wire vocabularies and the write path is what varies.
  - Organization grew _jobs_manager_class / _inference_api_manager_class so
    the shared class can hand out version-appropriate sub-managers.

Tests pin the invariant: TestV1IsDeletable AST-scans every management/v2/
module for v1 imports and re-imports them all behind a sys.meta_path blocker
that fails if any import chain reaches management.v1. Both still fail --
v2/workspace.py and v2/export.py remain to be written fresh -- which is the
point of landing them now.

TestFactoriesAreNotDuplicated replaces TestV1FactoryRoutesByVersion: the
manage_* factories are version-neutral dispatchers, so they belong in exactly
one place instead of being copied into v1/.
