+++ singlestoredb/management/v1/workspace.py
@@ -19,7 +19,13 @@ from typing import Union
 from ... import config
 from ... import connection
 from ...exceptions import ManagementError
+from ..billing import Billing as Billing
 from ..manager import Manager
+from ..organization import Organization
+from ..organization import Organizations as Organizations
+from ..region import Region
+from ..stage import Stage as Stage
+from ..stage import StageObject as StageObject
 from ..utils import camel_to_snake_dict
 from ..utils import ensure_within
 from ..utils import from_datetime
@@ -33,15 +39,6 @@ from ..utils import to_datetime
 from ..utils import ttl_property
 from ..utils import vars_to_str
 from ..versioned import VersionedMixin
-from .billing_usage import BillingUsageItem
-from .files import FileLocation
-from .files import FilesObject
-from .files import FilesObjectBytesReader
-from .files import FilesObjectBytesWriter
-from .files import FilesObjectTextReader
-from .files import FilesObjectTextWriter
-from .organization import Organization
-from .region import Region
 
 
 def get_organization() -> Organization:
@@ -91,704 +88,6 @@ def get_workspace(
             os.environ['SINGLESTOREDB_WORKSPACE']
         ]
     raise RuntimeError('no workspace group specified')
-
-
-class Stage(FileLocation):
-    """
-    Stage manager.
-
-    This object is not instantiated directly.
-    It is returned by ``WorkspaceGroup.stage`` or ``StarterWorkspace.stage``.
-
-    """
-
-    def __init__(self, deployment_id: str, manager: WorkspaceManager):
-        self._deployment_id = deployment_id
-        self._manager = manager
-
-    def open(
-        self,
-        stage_path: PathLike,
-        mode: str = 'r',
-        encoding: Optional[str] = None,
-    ) -> Union[io.StringIO, io.BytesIO]:
-        """
-        Open a Stage path for reading or writing.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            The stage path to read / write
-        mode : str, optional
-            The read / write mode. The following modes are supported:
-                * 'r' open for reading (default)
-                * 'w' open for writing, truncating the file first
-                * 'x' create a new file and open it for writing
-            The data type can be specified by adding one of the following:
-                * 'b' binary mode
-                * 't' text mode (default)
-        encoding : str, optional
-            The string encoding to use for text
-
-        Returns
-        -------
-        FilesObjectBytesReader - 'rb' or 'b' mode
-        FilesObjectBytesWriter - 'wb' or 'xb' mode
-        FilesObjectTextReader - 'r' or 'rt' mode
-        FilesObjectTextWriter - 'w', 'x', 'wt' or 'xt' mode
-
-        """
-        if '+' in mode or 'a' in mode:
-            raise ValueError('modifying an existing stage file is not supported')
-
-        if 'w' in mode or 'x' in mode:
-            exists = self.exists(stage_path)
-            if exists:
-                if 'x' in mode:
-                    raise FileExistsError(f'stage path already exists: {stage_path}')
-                self.remove(stage_path)
-            if 'b' in mode:
-                return FilesObjectBytesWriter(b'', self, stage_path)
-            return FilesObjectTextWriter('', self, stage_path)
-
-        if 'r' in mode:
-            content = self.download_file(stage_path)
-            if isinstance(content, bytes):
-                if 'b' in mode:
-                    return FilesObjectBytesReader(content)
-                encoding = 'utf-8' if encoding is None else encoding
-                return FilesObjectTextReader(content.decode(encoding))
-
-            if isinstance(content, str):
-                return FilesObjectTextReader(content)
-
-            raise ValueError(f'unrecognized file content type: {type(content)}')
-
-        raise ValueError(f'must have one of create/read/write mode specified: {mode}')
-
-    def upload_file(
-        self,
-        local_path: Union[PathLike, io.IOBase],
-        stage_path: PathLike,
-        *,
-        overwrite: bool = False,
-    ) -> FilesObject:
-        """
-        Upload a local file.
-
-        Parameters
-        ----------
-        local_path : Path or str or file-like
-            Path to the local file or an open file object
-        stage_path : Path or str
-            Path to the stage file
-        overwrite : bool, optional
-            Should the ``stage_path`` be overwritten if it exists already?
-
-        """
-        if isinstance(local_path, io.IOBase):
-            pass
-        elif not os.path.isfile(local_path):
-            raise IsADirectoryError(f'local path is not a file: {local_path}')
-
-        if self.exists(stage_path):
-            if not overwrite:
-                raise OSError(f'stage path already exists: {stage_path}')
-
-            self.remove(stage_path)
-
-        if isinstance(local_path, io.IOBase):
-            return self._upload(local_path, stage_path, overwrite=overwrite)
-
-        return self._upload(open(local_path, 'rb'), stage_path, overwrite=overwrite)
-
-    def upload_folder(
-        self,
-        local_path: PathLike,
-        stage_path: PathLike,
-        *,
-        overwrite: bool = False,
-        recursive: bool = True,
-        include_root: bool = False,
-        ignore: Optional[Union[PathLike, List[PathLike]]] = None,
-    ) -> FilesObject:
-        """
-        Upload a folder recursively.
-
-        Only the contents of the folder are uploaded. To include the
-        folder name itself in the target path use ``include_root=True``.
-
-        Parameters
-        ----------
-        local_path : Path or str
-            Local directory to upload
-        stage_path : Path or str
-            Path of stage folder to upload to
-        overwrite : bool, optional
-            If a file already exists, should it be overwritten?
-        recursive : bool, optional
-            Should nested folders be uploaded?
-        include_root : bool, optional
-            Should the local root folder itself be uploaded as the top folder?
-        ignore : Path or str or List[Path] or List[str], optional
-            Glob patterns of files or folders to ignore, for example,
-            ``**/*.pyc`` will ignore all ``*.pyc`` files in the directory
-            tree, and ``**/__pycache__`` will ignore those folders entirely.
-            Relative patterns are resolved against ``local_path``.
-
-        """
-        if not os.path.isdir(local_path):
-            raise NotADirectoryError(f'local path is not a directory: {local_path}')
-
-        stage_prefix = normalize_remote_path(stage_path)
-
-        if self.exists(stage_prefix) and not self.is_dir(stage_prefix):
-            raise NotADirectoryError(f'stage path is not a directory: {stage_path}')
-
-        ignore_files = resolve_ignore_files(local_path, ignore)
-
-        local_root = os.path.normpath(str(local_path))
-        root_name = os.path.basename(local_root)
-
-        for dir_path, dirs, files in os.walk(local_root):
-            if ignore_files:
-                # Prune ignored folders so their contents are skipped too
-                dirs[:] = [
-                    d for d in dirs
-                    if os.path.normpath(os.path.join(dir_path, d))
-                    not in ignore_files
-                ]
-            for fname in files:
-                # Normalized so it compares equal to the normalized
-                # glob results in ignore_files (e.g. local_path='.')
-                local_file_path = os.path.normpath(os.path.join(dir_path, fname))
-                if ignore_files and local_file_path in ignore_files:
-                    continue
-                rel = os.path.relpath(local_file_path, local_root)
-                if include_root:
-                    rel = os.path.join(root_name, rel)
-                # Remote paths always use '/', whatever the local platform
-                rel = rel.replace(os.sep, '/')
-                target = f'{stage_prefix}/{rel}' if stage_prefix else rel
-                self.upload_file(local_file_path, target, overwrite=overwrite)
-            if not recursive:
-                break
-
-        return self.info(stage_prefix)
-
-    def _upload(
-        self,
-        content: Union[str, bytes, io.IOBase],
-        stage_path: PathLike,
-        *,
-        overwrite: bool = False,
-    ) -> FilesObject:
-        """
-        Upload content to a stage file.
-
-        Parameters
-        ----------
-        content : str or bytes or file-like
-            Content to upload to stage
-        stage_path : Path or str
-            Path to the stage file
-        overwrite : bool, optional
-            Should the ``stage_path`` be overwritten if it exists already?
-
-        """
-        if self.exists(stage_path):
-            if not overwrite:
-                raise OSError(f'stage path already exists: {stage_path}')
-            self.remove(stage_path)
-
-        self._manager._put(
-            f'stage/{self._deployment_id}/fs/{stage_path}',
-            files={'file': content},
-            headers={'Content-Type': None},
-        )
-
-        return self.info(stage_path)
-
-    def mkdir(self, stage_path: PathLike, overwrite: bool = False) -> FilesObject:
-        """
-        Make a directory in the stage.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path of the folder to create
-        overwrite : bool, optional
-            Should the stage path be overwritten if it exists already?
-
-        Returns
-        -------
-        FilesObject
-
-        """
-        stage_path = re.sub(r'/*$', r'', str(stage_path)) + '/'
-
-        if self.exists(stage_path):
-            if not overwrite:
-                return self.info(stage_path)
-
-            self.remove(stage_path)
-
-        self._manager._put(
-            f'stage/{self._deployment_id}/fs/{stage_path}?isFile=false',
-        )
-
-        return self.info(stage_path)
-
-    mkdirs = mkdir
-
-    def rename(
-        self,
-        old_path: PathLike,
-        new_path: PathLike,
-        *,
-        overwrite: bool = False,
-    ) -> FilesObject:
-        """
-        Move the stage file to a new location.
-
-        Paraemeters
-        -----------
-        old_path : Path or str
-            Original location of the path
-        new_path : Path or str
-            New location of the path
-        overwrite : bool, optional
-            Should the ``new_path`` be overwritten if it exists already?
-
-        """
-        if not self.exists(old_path):
-            raise OSError(f'stage path does not exist: {old_path}')
-
-        if self.exists(new_path):
-            if not overwrite:
-                raise OSError(f'stage path already exists: {new_path}')
-
-            if str(old_path).endswith('/') and not str(new_path).endswith('/'):
-                raise OSError('original and new paths are not the same type')
-
-            if str(new_path).endswith('/'):
-                self.removedirs(new_path)
-            else:
-                self.remove(new_path)
-
-        self._manager._patch(
-            f'stage/{self._deployment_id}/fs/{old_path}',
-            json=dict(newPath=new_path),
-        )
-
-        return self.info(new_path)
-
-    def info(self, stage_path: PathLike) -> FilesObject:
-        """
-        Return information about a stage location.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage location
-
-        Returns
-        -------
-        FilesObject
-
-        """
-        res = self._manager._get(
-            re.sub(r'/+$', r'/', f'stage/{self._deployment_id}/fs/{stage_path}'),
-            params=dict(metadata=1),
-        ).json()
-
-        return FilesObject.from_dict(res, self)
-
-    def exists(self, stage_path: PathLike) -> bool:
-        """
-        Does the given stage path exist?
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to stage object
-
-        Returns
-        -------
-        bool
-
-        """
-        try:
-            self.info(stage_path)
-            return True
-        except ManagementError as exc:
-            if exc.errno == 404:
-                return False
-            raise
-
-    def is_dir(self, stage_path: PathLike) -> bool:
-        """
-        Is the given stage path a directory?
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to stage object
-
-        Returns
-        -------
-        bool
-
-        """
-        try:
-            return self.info(stage_path).type == 'directory'
-        except ManagementError as exc:
-            if exc.errno == 404:
-                return False
-            raise
-
-    def is_file(self, stage_path: PathLike) -> bool:
-        """
-        Is the given stage path a file?
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to stage object
-
-        Returns
-        -------
-        bool
-
-        """
-        try:
-            return self.info(stage_path).type != 'directory'
-        except ManagementError as exc:
-            if exc.errno == 404:
-                return False
-            raise
-
-    def _listdir(
-        self, stage_path: PathLike, *,
-        recursive: bool = False,
-        return_objects: bool = False,
-    ) -> List[Union[str, 'FilesObject']]:
-        """
-        Return the names (or FilesObject instances) of files in a directory.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the folder in Stage
-        recursive : bool, optional
-            Should folders be listed recursively?
-        return_objects : bool, optional
-            If True, return list of FilesObject instances. Otherwise just paths.
-
-        """
-        from .files import FilesObject
-        res = self._manager._get(
-            re.sub(r'/+$', r'/', f'stage/{self._deployment_id}/fs/{stage_path}'),
-        ).json()
-        if recursive:
-            out: List[Union[str, FilesObject]] = []
-            for item in res['content'] or []:
-                if return_objects:
-                    out.append(FilesObject.from_dict(item, self))
-                else:
-                    out.append(item['path'])
-                if item['type'] == 'directory':
-                    out.extend(
-                        self._listdir(
-                            item['path'],
-                            recursive=recursive,
-                            return_objects=return_objects,
-                        ),
-                    )
-            return out
-        if return_objects:
-            return [
-                FilesObject.from_dict(x, self)
-                for x in res['content'] or []
-            ]
-        return [x['path'] for x in res['content'] or []]
-
-    @overload
-    def listdir(
-        self,
-        stage_path: PathLike = '/',
-        *,
-        recursive: bool = False,
-        return_objects: Literal[True],
-    ) -> List['FilesObject']:
-        ...
-
-    @overload
-    def listdir(
-        self,
-        stage_path: PathLike = '/',
-        *,
-        recursive: bool = False,
-        return_objects: Literal[False] = False,
-    ) -> List[str]:
-        ...
-
-    def listdir(
-        self,
-        stage_path: PathLike = '/',
-        *,
-        recursive: bool = False,
-        return_objects: bool = False,
-    ) -> Union[List[str], List['FilesObject']]:
-        """
-        List the files / folders at the given path.
-
-        Parameters
-        ----------
-        stage_path : Path or str, optional
-            Path to the stage location
-        recursive : bool, optional
-            If True, recursively list all files and folders
-        return_objects : bool, optional
-            If True, return list of FilesObject instances. Otherwise just paths.
-
-        Returns
-        -------
-        List[str] or List[FilesObject]
-
-        """
-        from .files import FilesObject
-        stage_path = normalize_remote_path(stage_path, strip_leading=True) + '/'
-
-        if self.is_dir(stage_path):
-            out = self._listdir(
-                stage_path,
-                recursive=recursive,
-                return_objects=return_objects,
-            )
-            if stage_path != '/':
-                stage_path_n = len(stage_path.split('/')) - 1
-                if return_objects:
-                    result: List[FilesObject] = []
-                    for item in out:
-                        if isinstance(item, FilesObject):
-                            rel = '/'.join(item.path.split('/')[stage_path_n:])
-                            item.path = rel
-                            result.append(item)
-                    return result
-                out = ['/'.join(str(x).split('/')[stage_path_n:]) for x in out]
-            if return_objects:
-                return cast(List[FilesObject], out)
-            return cast(List[str], out)
-
-        raise NotADirectoryError(f'stage path is not a directory: {stage_path}')
-
-    def download_file(
-        self,
-        stage_path: PathLike,
-        local_path: Optional[PathLike] = None,
-        *,
-        overwrite: bool = False,
-        encoding: Optional[str] = None,
-    ) -> Optional[Union[bytes, str]]:
-        """
-        Download the content of a stage path.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage file
-        local_path : Path or str
-            Path to local file target location
-        overwrite : bool, optional
-            Should an existing file be overwritten if it exists?
-        encoding : str, optional
-            Encoding used to convert the resulting data
-
-        Returns
-        -------
-        bytes or str - ``local_path`` is None
-        None - ``local_path`` is a Path or str
-
-        """
-        return self._download_file(
-            stage_path,
-            local_path=local_path,
-            overwrite=overwrite,
-            encoding=encoding,
-            _skip_dir_check=False,
-        )
-
-    def _download_file(
-        self,
-        stage_path: PathLike,
-        local_path: Optional[PathLike] = None,
-        *,
-        overwrite: bool = False,
-        encoding: Optional[str] = None,
-        _skip_dir_check: bool = False,
-    ) -> Optional[Union[bytes, str]]:
-        """
-        Internal method to download the content of a stage path.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage file
-        local_path : Path or str
-            Path to local file target location
-        overwrite : bool, optional
-            Should an existing file be overwritten if it exists?
-        encoding : str, optional
-            Encoding used to convert the resulting data
-        _skip_dir_check : bool, optional
-            Skip the remote directory check when the caller already knows
-            ``stage_path`` refers to a file (e.g. from a directory listing)
-
-        Returns
-        -------
-        bytes or str - ``local_path`` is None
-        None - ``local_path`` is a Path or str
-
-        """
-        if local_path is not None and not overwrite and os.path.exists(local_path):
-            raise OSError('target file already exists; use overwrite=True to replace')
-        if not _skip_dir_check and self.is_dir(stage_path):
-            raise IsADirectoryError(f'stage path is a directory: {stage_path}')
-
-        out = self._manager._get(
-            f'stage/{self._deployment_id}/fs/{stage_path}',
-        ).content
-
-        if local_path is not None:
-            with open(local_path, 'wb') as outfile:
-                outfile.write(out)
-            return None
-
-        if encoding:
-            return out.decode(encoding)
-
-        return out
-
-    def download_folder(
-        self,
-        stage_path: PathLike,
-        local_path: Optional[PathLike] = None,
-        *,
-        overwrite: bool = False,
-    ) -> None:
-        """
-        Download a Stage folder to a local directory.
-
-        The contents of ``stage_path`` are written into ``local_path``,
-        which is created as the destination folder.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage folder
-        local_path : Path or str, optional
-            Local directory to create and download into. Defaults to the
-            name of the ``stage_path`` folder in the current directory.
-        overwrite : bool, optional
-            Should an existing directory / files be overwritten if they exist?
-
-        """
-        # ``listdir`` returns paths relative to ``stage_path``, so the folder
-        # prefix has to be added back on before making any remote calls.
-        stage_prefix = normalize_remote_path(stage_path, strip_leading=True)
-
-        if local_path is None:
-            local_path = os.path.basename(stage_prefix)
-            if not local_path:
-                raise ValueError(
-                    'local_path must be specified when downloading '
-                    'the root folder',
-                )
-
-        if not overwrite and os.path.exists(local_path):
-            raise OSError(
-                'target directory already exists; '
-                'use overwrite=True to replace',
-            )
-        if not self.is_dir(stage_prefix):
-            raise NotADirectoryError(f'stage path is not a directory: {stage_path}')
-
-        # Request objects so the file / directory type comes from the listing
-        # rather than an extra is_dir call per entry.
-        for entry in self.listdir(stage_prefix, recursive=True, return_objects=True):
-            rel_path = entry.path
-            target = ensure_within(local_path, os.path.join(local_path, rel_path))
-            if entry.type == 'directory':
-                os.makedirs(target, exist_ok=True)
-                continue
-            remote_path = (
-                f'{stage_prefix}/{rel_path}' if stage_prefix else rel_path
-            )
-            os.makedirs(os.path.dirname(target) or '.', exist_ok=True)
-            self._download_file(
-                remote_path, target,
-                overwrite=overwrite, _skip_dir_check=True,
-            )
-
-    def remove(self, stage_path: PathLike) -> None:
-        """
-        Delete a stage location.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage location
-
-        """
-        if self.is_dir(stage_path):
-            raise IsADirectoryError(
-                'stage path is a directory, '
-                f'use rmdir or removedirs: {stage_path}',
-            )
-
-        self._manager._delete(f'stage/{self._deployment_id}/fs/{stage_path}')
-
-    def removedirs(self, stage_path: PathLike) -> None:
-        """
-        Delete a stage folder recursively.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage location
-
-        """
-        stage_path = re.sub(r'/*$', r'', str(stage_path)) + '/'
-        self._manager._delete(f'stage/{self._deployment_id}/fs/{stage_path}')
-
-    def rmdir(self, stage_path: PathLike) -> None:
-        """
-        Delete a stage folder.
-
-        Parameters
-        ----------
-        stage_path : Path or str
-            Path to the stage location
-
-        """
-        stage_path = re.sub(r'/*$', r'', str(stage_path)) + '/'
-
-        if self.listdir(stage_path):
-            raise OSError(f'stage folder is not empty, use removedirs: {stage_path}')
-
-        self._manager._delete(f'stage/{self._deployment_id}/fs/{stage_path}')
-
-    def __str__(self) -> str:
-        """Return string representation."""
-        return vars_to_str(self)
-
-    def __repr__(self) -> str:
-        """Return string representation."""
-        return str(self)
-
-
-StageObject = FilesObject  # alias for backward compatibility
 
 
 class Workspace(VersionedMixin):
@@ -1587,6 +886,12 @@ class StarterWorkspace(VersionedMixin):
 
     """
 
+    #: Base management API path for the shared-tier resource. v2 renamed this
+    #: from ``virtualWorkspaces`` to ``virtualClusters``; every shared-tier
+    #: request is built from this attribute so the rename is a one-line
+    #: override in the v2 subclass.
+    _sharedtier_path = 'sharedtier/virtualWorkspaces'
+
     name: str
     id: str
     database_name: str
@@ -1701,7 +1006,7 @@ class StarterWorkspace(VersionedMixin):
             raise ManagementError(
                 msg='No workspace manager is associated with this object.',
             )
-        self._manager._delete(f'sharedtier/virtualWorkspaces/{self.id}')
+        self._manager._delete(f'{self._sharedtier_path}/{self.id}')
 
     def refresh(self) -> StarterWorkspace:
         """Update the object to the current state."""
@@ -1741,9 +1046,9 @@ class StarterWorkspace(VersionedMixin):
             raise ManagementError(
                 msg='No workspace manager is associated with this object.',
             )
-        res = self._manager._get('sharedtier/virtualWorkspaces')
+        res = self._manager._get(self._sharedtier_path)
         return NamedList(
-            [StarterWorkspace.from_dict(item, self._manager) for item in res.json()],
+            [type(self).from_dict(item, self._manager) for item in res.json()],
         )
 
     def create_user(
@@ -1784,7 +1089,7 @@ class StarterWorkspace(VersionedMixin):
             payload['password'] = password
 
         res = self._manager._post(
-            f'sharedtier/virtualWorkspaces/{self.id}/users',
+            f'{self._sharedtier_path}/{self.id}/users',
             json=payload,
         )
 
@@ -1803,77 +1108,6 @@ class StarterWorkspace(VersionedMixin):
             'user_id': user_id,
             'password': returned_password,
         }
-
-
-class Billing(object):
-    """Billing information."""
-
-    COMPUTE_CREDIT = 'compute_credit'
-    STORAGE_AVG_BYTE = 'storage_avg_byte'
-
-    HOUR = 'hour'
-    DAY = 'day'
-    MONTH = 'month'
-
-    def __init__(self, manager: Manager):
-        self._manager = manager
-
-    def usage(
-        self,
-        start_time: datetime.datetime,
-        end_time: datetime.datetime,
-        metric: Optional[str] = None,
-        aggregate_by: Optional[str] = None,
-    ) -> List[BillingUsageItem]:
-        """
-        Get usage information.
-
-        Parameters
-        ----------
-        start_time : datetime.datetime
-            Start time for usage interval
-        end_time : datetime.datetime
-            End time for usage interval
-        metric : str, optional
-            Possible metrics are ``mgr.billing.COMPUTE_CREDIT`` and
-            ``mgr.billing.STORAGE_AVG_BYTE`` (default is all)
-        aggregate_by : str, optional
-            Aggregate type used to group usage: ``mgr.billing.HOUR``,
-            ``mgr.billing.DAY``, or ``mgr.billing.MONTH``
-
-        Returns
-        -------
-        List[BillingUsage]
-
-        """
-        res = self._manager._get(
-            'billing/usage',
-            params={
-                k: v for k, v in dict(
-                    metric=snake_to_camel(metric),
-                    startTime=from_datetime(start_time),
-                    endTime=from_datetime(end_time),
-                    aggregateBy=aggregate_by.lower() if aggregate_by else None,
-                ).items() if v is not None
-            },
-        )
-        return [
-            BillingUsageItem.from_dict(x, self._manager)
-            for x in res.json()['billingUsage']
-        ]
-
-
-class Organizations(object):
-    """Organizations."""
-
-    def __init__(self, manager: Manager):
-        self._manager = manager
-
-    @property
-    def current(self) -> Organization:
-        """Get current organization."""
-        res = self._manager._get('organizations/current').json()
-        return Organization.from_dict(res, self._manager)
 
 
 class WorkspaceManager(Manager):
@@ -1907,6 +1141,10 @@ class WorkspaceManager(Manager):
     #: Object type
     obj_type = 'workspace'
 
+    #: Base management API path for the shared-tier resource. See
+    #: :attr:`StarterWorkspace._sharedtier_path`.
+    _sharedtier_path = 'sharedtier/virtualWorkspaces'
+
     @property
     def workspace_groups(self) -> NamedList[WorkspaceGroup]:
         """Return a list of available workspace groups."""
@@ -1916,7 +1154,7 @@ class WorkspaceManager(Manager):
     @property
     def starter_workspaces(self) -> NamedList[StarterWorkspace]:
         """Return a list of available starter workspaces."""
-        res = self._get('sharedtier/virtualWorkspaces')
+        res = self._get(self._sharedtier_path)
         return NamedList([StarterWorkspace.from_dict(item, self) for item in res.json()])
 
     @property
@@ -2190,7 +1428,7 @@ class WorkspaceManager(Manager):
         :class:`StarterWorkspace`
 
         """
-        res = self._get(f'sharedtier/virtualWorkspaces/{id}')
+        res = self._get(f'{self._sharedtier_path}/{id}')
         return StarterWorkspace.from_dict(res.json(), manager=self)
 
     def create_starter_workspace(
@@ -2231,12 +1469,12 @@ class WorkspaceManager(Manager):
         if project_id is not None:
             payload['projectID'] = project_id
 
-        res = self._post('sharedtier/virtualWorkspaces', json=payload)
+        res = self._post(self._sharedtier_path, json=payload)
         virtual_workspace_id = res.json().get('virtualWorkspaceID')
         if not virtual_workspace_id:
             raise ManagementError(msg='No virtualWorkspaceID returned from API')
 
-        res = self._get(f'sharedtier/virtualWorkspaces/{virtual_workspace_id}')
+        res = self._get(f'{self._sharedtier_path}/{virtual_workspace_id}')
         return StarterWorkspace.from_dict(res.json(), self)
