+++ singlestoredb/management/v2/billing_usage.py
@@ -1,4 +1,10 @@
 #!/usr/bin/env python
-"""SingleStoreDB Billing Usage API v2."""
-from ..v1.billing_usage import BillingUsageItem as BillingUsageItem
-from ..v1.billing_usage import UsageItem as UsageItem
+"""
+SingleStoreDB Billing Usage API v2.
+
+``GET /v1/billing/usage`` and ``GET /v2/billing/usage`` are identical, so the
+implementation lives in the shared
+:mod:`singlestoredb.management.billing_usage` module.
+"""
+from ..billing_usage import BillingUsageItem as BillingUsageItem
+from ..billing_usage import UsageItem as UsageItem
